refactor: move searchParamsToObject to server-utils and update imports
