+++ src/rpc/server/server-utils.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from "vitest";
-import { searchParamsToObject } from "./search-params-to-object";
+import { searchParamsToObject } from "./server-utils";
 
 describe("searchParamsToObject", () => {
   it("should retrieve a single query parameter", () => {
